+++ src/layer/layer.h
@@ -32,7 +32,7 @@ class Layer {
         Matrix forward(const Matrix& x);
         Matrix backward(const Matrix& x, const vector<int>& y_hat); // softmax layer
         Matrix backward(const Matrix& x, const Matrix& dh); // hidden layer
-        void update(double lr);
+        void update(float lr);
         void setTrain(bool train);
         Matrix getH() const { return h; }
 };
